Only delete import-dirs, not library-dirs
If cabal installs a package with the `--extra-library-dirs` flag, then
`ghc-pkg` will report those libraries when invoking `ghc-pkg field
<package> library-dirs`, causing `cab` to delete them. This can be
catastrophic if among those directories are shared locations for library
files (e.g., MinGW libraries on Windows).

To work around this, we note that Cabal almost always installs all `.hi`
and `.p-hi` files to a single directory, which it places in the
`import-dirs` field. Furthermore, we note that in practice, Cabal seems to
make the set of `import-dirs` a subset of the set of `library-dirs`, so to
avoid deleting directories that aren't Cabal-related, we can simply delete
`import-dirs` alone.